ES-977 Allow Kafka TLS with no truststore
